requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pyspark==3.5.6
apache-airflow-providers-apache-spark==4.8.0
apache-airflow-providers-common-compat==1.6.0
apache-airflow-providers-google==16.1.0
apache-airflow-providers-microsoft-mssql==4.3.2
pyarrow==20.0.0
google-cloud-bigquery==3.35.1
pyodbc==5.2.0